[WIP] Allow multiple `should_capture` filters for Sentry
Sentry's Raven allows configuration of a method that determines
whether to capture errors or not: `should_capture=`. However, this
method only accepts a single `Proc` as an argument, and even then we
are already using this method in this gem in order to hook into Statsd.

In order to allow clients to use the `should_capture` method, this
change adds a `GovukError::Configuration` class, which sets up an array
of `Proc`s, which must _all_ return `true` to capture the error.