Add Unicorn as a dependency
All GOV.UK apps use unicorn as the web server, but we're using a wild
variety of versions.

This PR adds the gem as a dependency so we'll be using the same
version. This will make debugging issues easier, as well as switching
to a new web server like Puma.