Setup logging through logstasher
We currently use 2 types of logging on GOV.UK. One that sends Rails'
request logs as JSON to a file (`production.json.log`) and a newer one
that sends the logs to STDOUT. The latter is the correct 12-factor
method of logging things (https://12factor.net/logs).

This adds the standard config for logging with logstasher to this gem,
and configures the process so that only JSON is sent to STDOUT and the
rest to STDERR. This method is already used in government-frontend,
publishing-api and a few other apps. I found this a bit dark magic so
I've added a bunch of comments to explain what's happening.