Configuration for Unicorn
This is to provide a shared configuration for unicorn. Currently GOV.UK
apps rely on this notification to be provided by puppet through
[/etc/govuk/unicorn.rb][1] being created on each host. Then at the point
of running an app there is a [check][2] for whether a `config/unicorn.rb`
exists otherwise the puppet file will be used.

That approach has a few downsides:

1) There is code injected into the app at runtime which is not run in
other environments, notably test ones
2) There is a coupling between apps and govuk-puppet which is
unnecessary
3) It's something of an anomaly in govuk_app_config to provide a tool
(unicorn) but not it's configuration.

This approach resolves those but does have it's own downside that there
does not seem to be a way to configure unicorn automatically so apps are
required to have a config/unicorn.rb file to require the configuration.

[1]:
https://github.com/alphagov/govuk-puppet/blob/d8ca60213355704975ab655e4549069d729f15bf/modules/govuk/files/etc/govuk/unicorn.rb
[2]:
https://github.com/alphagov/govuk-puppet/blob/a6c51d887a6501f02766b7279127b60f02037a7f/modules/govuk/files/usr/local/bin/govuk_spinup#L50-L53