Move logging configuration outside of global
Having this in the global space means it can be run before applications
have started up and cause surprises with STDOUT configuration.

The catalyst for moving this was discovering that it was breaking
unicorn write to stdout configuration and was causing unicorn to output
to actual STDOUT.

Since this logging only actually takes effect if we use this module
(which is only started automatically in Rails apps) we probably
shouldn't have it causing side effects.