Configure the GDS API Adapters logger if available
The GDS API Adapters gem can log useful information about the requests
being made, but by default, it's configured to use a null logger for
this.

Instead, use the govuk_app_config gem to change the default
configuration for GDS API Adapters to use the logstasher logger.

This should make this logging from the GDS API Adapters more
available, and simplify the configuration in apps.