Enable initialized Healcheck classes
Adds support for adding initialized classes for the health check chain:

```ruby
get "/healthcheck", to: GovukHealthcheck.rack_response(
  GovukHealthcheck::SidekiqRedis,
  GovukHealthcheck::ActiveRecord,
  GovukHealthcheck::InitializedCheck.new(:param),
)
```

Enable applications to reuse the same Check without being forced
to use inheritance or other mechanisms.

This is introduced in[1], and is part of PR[2]. We aim to reuse the same check for multiple metrics.

[1]: https://github.com/alphagov/content-performance-manager/pull/1131/commits/73aec09c289f785c06040845662c61a7f4d163e1
[2]:  https://github.com/alphagov/content-performance-manager/pull/1131