Make ActiveRecord healthcheck less flakey
Previously the AR healthcheck used to call '.connected?' and return a
status based on the boolean return value. But if the app is not being
used much, then the connection pool may be empty as all the idle
connections will have been removed. This changes the approach to
explicitly requests a connection if it doesn't already exist, which
means we should have less false negatives for low-traffic apps.