Use Rails DSL to define content security policy
This refactors the CSP to use the Rails DSL, this is backwards
compatible with the previous implementation. This refactor provides a few
benefits:

- no longer needs to set up the actual policy, just configuration
  needed;
- apps can add/remove things to the policy if needed;

I've reduced the amount of comments in this file for brevity. A lot of
the example URLs felt unnecessary for the source code and felt more like they
belonged in a commit message.