Send Rails.logger calls to stdout as JSON
This changes the approach to logging so that calls to Rails.logger are
included in the stdout of the app and are included in Kibana. Previously
they were sent to stdout. This brings the logging to
be consistent with the approach that is used for sidekiq and with the
comment at the top of govuk_logging file.

This works by creating a new formatter that creates a JSON line that is
output to stdout.

It has a couple of things that make it not perfect:

1) When using with `rails server` log output is outputted twice. Rails
has special code to stop duplicate output to stdout [1], however the
redirect we make with $stdout to $stderr breaks this. This PR doesn't
introduce this problem, but was something I witnessed in testing.
2) Tagged logging updates the message and can't be used in JSON payload.
When testing this with Content Publisher I noticed it was logging a
@message of "[1c52e5f4-6e1d-41dc-a1cd-90a00a468f54] testing logging",
where the UUID is a request_id. I hoped that this is something that
could be a different JSON field however the Rails tagged logging only
has values and not keys [2]

[1]: https://github.com/rails/rails/commit/3d10d9d6c3b831fe9632c43a0ffec46104f912a7
[2]: https://github.com/rails/rails/blob/bf19b8774e20e98f7fdcd3ac82ee17f9adee22d8/railties/lib/rails/rack/logger.rb#L26