Fix incorrect error constant
There is no such constant in gds-api-adapters as
GdsApi::HTTPIntermittent. Instead there are two different constants that
are prefixed with GdsApi::HTTPIntermittent:
GdsApi::HTTPIntermittentClientError and
GdsApi::HTTPIntermittentServerError. Thus errors are now being reported
that would previously been excluded.

Presumably this mistake was made as we were previously just checking the
prefix of the class name in [1].

I decided we should only exclude intermittent server errors as applications
should be informed if they are making a client error (the case where you
would get a HTTPIntermittentClientError would be for too many requests, 429).

[1]: https://github.com/alphagov/govuk_app_config/commit/3c46b628a899717e7aec40273cf920bf7ff1a7c1