Revert rails_report_rescued_exceptions = false
It turns out that this setting was more aggressive than we anticipated
and was actually preventing the reporting of any exception that Rails
handled. We had thought that this only prevented the reporting of
exceptions specially configured as Rails rescue responses.

It transpires that this stops the reporting of any exceptions that are
handled by ActionDispatch::ShowExceptions [1] which is effectively all
Rails exceptions as per raven-ruby [2].

[1]: https://github.com/rails/rails/blob/38ec8db3f7d3c79936cdaa0a136b26ee1c977f7d/actionpack/lib/action_dispatch/middleware/show_exceptions.rb
[2]: https://github.com/getsentry/raven-ruby/blob/d9d41784c3b5888295bf4b013416f132045b22d4/lib/raven/integrations/rails.rb#L51-L52