Add all default Rails rescue response exceptions to ignore list
This adds in all the Rails default exceptions as is listed in:
https://github.com/rails/rails/blob/38ec8db3f7d3c79936cdaa0a136b26ee1c977f7d/actionpack/lib/action_dispatch/middleware/exception_wrapper.rb
and removes a legacy exception that was removed in Rails 4.

I did consider doing something like:

```
if defined?(ActionDispatch)
  config.excluded_exceptions += ActionDispatch::ExceptionWrapper.rescue_responses.keys
end
```

but I thought it was a bit opaque, particularly because the
rescue_responses attribute is edited during Rails initialisation, so
could be hard to understand why this had a particular setting.