Begin writing tests for GovukError.configure
In the next few commits, we're going to be pushing
`GovukError.configure` further. There is no configurable way of
hooking into the 'should_capture' callback at the govuk_app_config
level without it being overwritten at the app level.

We'll be building behaviour to do just that, so let's set up the
beginning of a test suite. I've written some tests for the
'silence_ready' logic, which is conditional based on the state
of the environment. For that reason, it's worth testing (as
opposed to a lot of config which is just string-based; tests
would not be appropriate for these).

The commit that added `silence_ready` in the first place explains:
fad652b23d8691797fa4994bb7a27191d2481fad
> Stop saying errors are not going to be captured in dev
> This is useful to have logged in production, but is distracting
everywhere else.

This is now (slightly) better documented as a result of the
tests.