Read data sync times from GOVUK_DATA_SYNC_PERIOD ENV variable
We will do some work in govuk-puppet to create this ENV variable
based on the time ranges that are already defined in puppet:
https://github.com/alphagov/govuk-puppet/blob/4bcf837e0dc4f48be59716921841307dfa291ebe/modules/govuk_data_sync_in_progress/manifests/init.pp#L20-L24

I have encapsulated the time parsing in its own class:
GovukDataSync. This sticks to the sole responsibility principle.

It is now sufficient for us to test that DefaultConfiguration
calls the GovukDataSync class and that it relies on its
'.in_progress?' method in determining whether or not we are in the
data sync period. Now our default_configuration tests test just
the business logic - that PostgreSQL errors should be ignored
during the sync, and nothing else - and not any extra time parsing
logic.

A beautiful side-effect of the test:
"passes the 'GOVUK_DATA_SYNC_PERIOD' ENV variable to GovukDataSync"
...is that the test will fail if you attempt to initialise
GovukDataSync inside the should_capture callback. We can therefore
be sure that the ENV variable is only being used at startup and
should not run into any quirky runtime errors.
This is a deliberate decision, so I have appended ", at startup"
to that test's description.

NB: As this is a lambda, we cannot use a `return` statement to
exit the `should_capture` lambda early. We have to process all
the logic up front and then use ruby's implicit return.
https://www.reddit.com/r/ruby/comments/4z86sg/another_lambda_vs_proc_vs_blocks_article/