Add healthcheck for remote Rails cache
We use a local Memcached instance for the cache store in several of
our apps [1] [2]. If the cache is unavailable, it will print error
logs but an exception won't be thrown. While the behaviour is graceful,
it's still a problem that the caching we rely on is not working as
expected, so we should surface this in healthchecks.

[1]: https://github.com/alphagov/collections/blob/a08857d95771e813338fbbe8542a8f5b536ea6d4/config/environments/production.rb#L58
[2]: https://github.com/alphagov/whitehall/blob/774eb78cd4c75622e786bdb18762850cd0b91840/config/environments/production.rb#L61