Also ignore subclasses of `data_sync_excluded_exceptions`
We're seeing PG:UndefinedTable errors in Sentry, which we'd like
to ignore during the data sync. By default we're ignoring PG::Error
exceptions already, but this is a specific sub class of PG::Error
(and there are many more we envisage wanting to ignore). We
could list all of these subclasses in the
`data_sync_excluded_exceptions` array, but what we'd rather do
is describe the parent class and have any of its children
exceptions ignored.

`PG::Error === PG::UndefinedTable.new` is true, whereas
`"PG::Error" === PG::UndefinedTable.new.class` is false.
This commit makes the default `should_capture` behaviour dig
deeper into the exception, checking that the exception type
is classed as any one of the errors we've chosen to ignore.

From now on, any exceptions that are a subclass of `PG::Error`
will be ignored during the data sync.

https://trello.com/c/rAlvGlSp/2123-5-ignore-data-sync-errors-in-govukappconfig