Ignore errors that occur in temporary environments
We sometimes spin up temporary environments (e.g. 'test-pink-aws')
while developing applications. If there are errors in those apps,
we can accidentally send many thousands of errors to Sentry, which
can mean we hit rate limits on our other, production,
environments, as we have a single shared Sentry account.

This commit introduces a `active_sentry_environments` config
property, which is checked against the SENTRY_CURRENT_ENV ENV
variable. If an error occurs in an environment we care about,
it is reported to Sentry, otherwise it gets ignored.

There is prior art for this - in #160 we added a similar check,
such that if a particular error occurs during the nightly data
sync, it gets ignored. I've made sure this new functionality is
compatible with what's already been added.

Trello: https://trello.com/c/GfFwJG0K/2180-investigate-disabling-sentry-on-environments-we-dont-care-about-timebox-2-days