Merge should_capture into before_send
This is necessary because `should_capture` is deprecated, but requires
slight behaviour changes in order to meet the expectations of both —
`should_capture` returns a boolean, but `before_send` returns nil if the
error is to be dropped; we want to support both the previous
`should_capture` behaviour (filtering) and the existing `before_send`
usecase of incrementing statsd counters.

In the process we also want to ensure that newly-added callbacks are all
run in order, that the statsd counters are incremented once and only
once, and that a callback returning nil prevents the final callback from
incrementing the counters.

`before_send` filtering behaviour is documented at
https://docs.sentry.io/platforms/ruby/configuration/filtering/