Allow configuration_spec.rb to be run in isolation
Running `bundle exec rspec spec/govuk_error/configuration_spec.rb`
resulted in errors where `GovukStatsd` wasn't defined. I've now
moved this stub to the top of the test file and indicated that it
has a `increment` method.

I've also fixed the return value of `nil` in configuration.rb,
since this value is only used on line 45 as a boolean, so it's
good that we return an explicit `false`. (The return value of `nil`
is what we want from the _lambda_ return values, but not this
internal bit of logic).