Fix `before_send` lambdas not being called
In #196, we changed `should_capture` to `before_send` in preparation
for [the removal of the `should_capture` method](https://github.com/getsentry/sentry-ruby/blob/master/MIGRATION.md#removed)
in sentry-ruby.

Because we already had some `before_send` logic, and it wasn't
idempotent (i.e. it must only be evaluated once, as it increments
counters and we want to avoid double-counting), we had to
re-architect how the default behaviour is merged with any custom
behaviour the downstream developer may set.

Since #196 was merged, we've seen errors occuring in Sentry, which
should be being ignored via the datasync logic. This suggests the
lambdas are not being evaluated.

I believe the removal of `before_send=` in configure.rb in
https://github.com/alphagov/govuk_app_config/pull/196/files#diff-5617fa12f2b7ce8393099a22ed6ae89a20e286273e17190f2bf9ec48ef5b46a3L2
has meant some of our setup code would not be reached, unless the
downstream app calls` before_send=` (and none of them do so far).

I also think the override of the `before_send =` on `super` may
be causing problems.

I've reverted to the previous architecture, whilst keeping the new
method names, to use an invocation of `super` that we know was
once working.

As a caveat, we've had to pass a harmless lambda in configure.rb,
to ensure the necessary setup code is called.