Use stub_const to define Rails::Application
This avoids defining a global class which looks like it's limited to
just the block, when actually it's not.

I've had to set `Rails.application` to `nil` after the test to ensure we
don't end up with the wrong Rails application being registered depending
on the ordering of the tests. At the moment, the `Rails.application`
variable holds state between tests.