Add test checking our default Sentry config
Attempting to pull govuk_app_config v4.0.0.pre.1 into a project
led to a failure, because our default Sentry configuration is
currently invalid. This will be fixed up in subsequent commits,
but to begin with, we should have a test that exercises our default
config and ensures that it doesn't raise any errors.