Remove `transport_failure_callback` config
In sentry-ruby, this raises an error:

```
NoMethodError: undefined method `transport_failure_callback=' for #<GovukError::Configuration:0x000055f6d3cd1368>
```

`transport_failure_callback` was removed and, in the migration
guide, advised to be swapped with `before_send`:
https://docs.sentry.io/platforms/ruby/migration/

However, there is no documentation around _how_ to replace this
with `before_send`, which is already rather a complicated method
as we've overloaded it with conditional-exception-ignoring logic.

I fear that the most straightforward option is to simply remove it.
It will mean making some tweaks to our Grafana dashboards in
govuk-puppet:
https://github.com/alphagov/govuk-puppet/search?q=error_reports_failed