Replace `active_sentry_environments` with `enabled_environments`
The custom option `active_sentry_environments` was added in
https://github.com/alphagov/govuk_app_config/pull/168.

However, it turns out there is already a native option for this:
`enabled_environments`. See
https://docs.sentry.io/platforms/ruby/configuration/options/

This allows us to simplify our code a fair bit.

I've also removed the stubbed `sending_allowed?` method as it was
interfering with some tests and giving a false sense of security.
We're now providing all of the necessary configuration for
the `Sentry::Configuration` instance to return `true` on its
`valid?` method, so no need for the hack.
Each config option is documented in an inline comment as otherwise
it can be quite obtuse.