Call `Sentry.init` via `GovukError.configure`
This is required for sentry-ruby to work properly. We used to
proxy our `configure` method through to `Sentry.configure`, but
this method is deprecated as of v4 of the gem.

Moreover, the `Sentry.init` method clears any previous configuration,
rather than layering additional configuration on top. In
subsequent commits, we will re-architect GovukError to force only
one `configure` call.