Revert changes introduced in 3.1.0 and 3.1.1
In 3.1.0/3.1.1 we merged the `should_capture` method into the
`before_send` method in preparation for swapping out `sentry-raven`
for `sentry-ruby`. However, it broke the "ignore certain errors
when they happen during the data sync" behaviour.

This is subsequently fixed in v4, but not all apps can/will be
upgraded to v4, so we should offer a stable v3.