Call original on Sentry methods to reveal issues
This configures the mocks for this method to call original so that in
any problems with the API can be exposed - as far as I can tell calling
these has no side effects.

The reason I'm adding this in is because these methods cause errors on
Ruby 3, but because we use mocks our test don't expose it. By calling
original these tests raise errors.