Add a warning for usage with Sidekiq without Sentry::Sidekiq
Since sentry-raven was discontinued in favour of sentry-ruby there is
now a separate gem for sentry-sidekiq error handling (this was
previously handled in sentry-raven).

Apps that don't have sentry-sidekiq installed will miss out on
exceptions generated by Sidekiq so it seems prudent to warn users.

I decided against adding sentry-sidekiq as a dependency of this repo as
this would add Sidekiq and all it's dependencies to every app that uses
it and I imagine only around 50% or less use Sidekiq.