Reorganise guidance on built-in checks
Having a list of checks is useful after the general description of
how to add them to an app.

The guidance about naming / modules was a bit vague, especially now
there are no abstract classes. It was unclear why some classes should
have a "Check" suffix and not others. I don't think it's a problem to
always have a  "Check" suffix.