Add GovukProxy::StaticProxy class
This class can be loaded as Rack middleware and will proxy any requests
with the path prefix of `/assets/static` to backend specified. This is
useful as it enables Static to use relative paths and the dependent
frontend application can forward those requests. This allows us to
switch Static to use relative path and allow us to have development
  environment and not run Static locally (i.e. forward request to
  production version of Static).