Avoid INFO logspew from the Sentry gem.
The Sentry gem spams stdout with INFO logs like this by default, which
is particularly annoying when troubleshooting. We don't need info-level
logs from the Sentry integration itself.

    [Transport] Sending envelope with items [sessions]  to Sentry

This change doesn't affect the logging of events to Sentry, only the
logging of messages from the Sentry gem itself. We're therefore only
interested in log messages which might indicate that something is wrong,
not log messages that simply indicate that something expected is
happening - hence WARN is the appropriate level here.