Fix an ownership bug when configuring the Sentry logger.
Sentry's default logger turns out to be a reference to Rails.logger by
default, despite the [sentry-rails docs] saying that it defaults to a new
Sentry::Logger. So I was unwittingly mutating Rails.logger through a
shared reference. Oops.

Kudos to super sharp-eyed @kevindew for catching this!

It turns out that changing the log level of a gem in Rails is hard, so
instead we fix the logspam by just disabling the Sentry feature that
causes the INFO messages. Turns out we weren't using that feature
anyway.

[sentry-rails docs]: https://docs.sentry.io/platforms/ruby/guides/rails/configuration/options/