Add plurals dropped from rails-i18n in 7.0.6
The 7.0.6 release of rails-i18n dropped support for a number of plurals
[1] and this causes our apps that make use of these languages to error
and thus be unable to update to rails-i18n 7.0.6

This commit ports over these rules from rails-i18n to maintain the
behaviour.

There is also something of a duplicate of this file in
[rails_translation_manager](https://github.com/alphagov/rails_translation_manager/blob/main/config/locales/plurals.rb)
that may also need an update. I'm unsure of the status of that file
though as it is already out-of-sync with this one and I think this one
takes priority.

[1]: https://github.com/svenfuchs/rails-i18n/pull/1017