Update Plek support to allow version 5
This updates the constraints on Plek so that version 5 can be used,
which is compatible with this gem.

In time we should drop support for v4 but this is currently unnecessary
as it remains compatible.

This also updates the syntax on the usage of Plek since it could use a
class method rather than doing an instantiation so can be a bit shorter.