Set default_src to be self in CSP
Only default to allow resources to be loaded from the same host as the
request.

My understanding is that https means that any https resource can be
loaded, which makes it a little unsafe. And the use of GOVUK_DOMAINS is
unnecessary as this is configured just for directives that need it.