Update CSP documentation
This updates the CSP with documentation on how you should edit it and
our current intentions of rules to adjust before making it live.

The great fear of this CSP is that once we get it out and it's secure,
devs start adding insecure or app specific things. This tries to explain
it somewhat, however I expect inline comments will also be needed.