Add YouTube image host to image_src policy
This is to support the upcoming YouTube card component [1] that makes
use of YouTube thumbnails.

Unlike most of the host directives in this file this one specifies the
scheme to be https. My understanding is that it was an earlier mistake
of ours not to specify schemes, since we expect any of these resources
to be https. At a later point we can migrate the others.

[1]: https://github.com/alphagov/govuk_publishing_components/pull/3156