Strip out sensitive information from error messages
There are some bits of infrastructure we can connect to via urls that contain usernames and passwords. If we misconfigure the URL, we can end up logging this information in the message of a 'bad uri' error.  This PR ensures that we do not send messages with these urls to sentry, avoiding leaking passwords (such as [1]). This commit adds this logic to a 'before_send' hook.

[1]: https://sentry.io/organizations/govuk/issues/3798011523/?project=202242