Replace underscore method with gsub
This was using the method `underscore` which is a rails method. However it was being called from places that did not have rails e.g. the cache clearing service resulting in an error.

So we have replaced the rails method with a ruby one that should be able to be run in more applications.