Remove the allow-list of Sentry environment names.
`active_sentry_environments` was a hard-coded list of allowable values
for the Sentry `environment` tag. It was introduced during the
Replatforming discovery phase in 2020 in order to prevent the Puppet/EC2
stack in the test account from sending events to Sentry.

(It was done this way because at that time, disabling Sentry in the
test environment was more difficult than it should have been, while some
teams had Sentry alerts and Slack subscriptions which didn't select by
environment).

We don't need this feature any more and we're now better off without it,
because it complicated the semantics of `SENTRY_CURRENT_ENV` in a way
that is surprising to the user.

If we want to run without Sentry in future, we can just not set
`SENTRY_DSN`.