Add HttpParserError to excluded exceptions
Sentry ruby 5.9 ensured that low level puma errors were captured [1] and propagated to sentry. This ended up in a lot of noise [2] in sentry when applications were bumped to this version of sentry ruby, particularly with Puma::HttpParser errors.

These errors are often unactionable and we believe a result of crawler activity. This suppresses these errors in order to reduce sentry noise.

1: https://github.com/getsentry/sentry-ruby/pull/2026
2: https://govuk.sentry.io/issues/4154286086/?project=202217