Move requires to top of file
This is needed because I want to use a class from
prometheus_exporter/metric as the default for a method parameter, which
means I need to require the prometheus code before I use it.

Previously these requires happened conditionally, so if
`should_configure` was false, the prometheus code wouldn't be loaded at
all. From a quick look at the prometheus exporter code, I don't think
the requires have any side effects, so it should be fine to load them
even when we're not configuring prometheus.