Allow gov.uk domains to embed pages
We initially added the strict and OWASP recommended 'none' directive based on
the assumption that only side-by-side-browser tool (retired in November 2022)
was preventing us from implementing it.

However some other internal GOV.UK apps use iframes:
- Search Admin
    - Best bets (queries)
    - External links (recommended-links)
- Content Publisher (Preview feature)

This policy will still ensure sufficient security yet will allow internal
GOV.UK domains to embed pages.  It's added to the global base policy because
given the number of frontend application it may be difficult to predict which
frontend app renders the page that we want to iframe. It will reduce the need
to apply a CSP modification in individual apps.